split status file dedup to a separate file - no code changes
