+++ src/daemon/signal-handler.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "common.h"
-#include "daemon/daemon-status-file.h"
+#include "daemon/status-file.h"
 
 #ifdef ENABLE_SENTRY
 #include "sentry-native/sentry-native.h"
